+++ src/utilities/functions.ts
@@ -2,10 +2,7 @@ import * as vscode from "vscode";
 import * as fs from "fs";
 import * as yaml from "yaml";
 import * as cp from "node:child_process";
-import * as util from "node:util";
 import * as c from "./constants";
-
-const exec = util.promisify(cp.exec);
 
 type PatternSet = [string, string[]];
 type PathDetails = {
@@ -22,6 +19,11 @@ type ExtendedTempFile = {
 	tempFile: vscode.Uri,
     originalFile: vscode.Uri,
 	content: string
+};
+
+type Answer = {
+	stdout:string,
+	stderr:string
 };
 
 export function executeInTerminal(commandArray:string[], terminal:vscode.Terminal)  {
@@ -47,35 +49,13 @@ export function dissectUri(file:vscode.Uri) : PathDetails {
 
 export function getTempUri(file:vscode.Uri) : vscode.Uri {
 	const fd = dissectUri(file);
-	const tempFileName = `${fd.filePureName}.${getTempFilePreExtension()}.${fd.extension}`;
+	const tempFileName = `${fd.filePureName}.${getSettingTempFilePreExtension()}.${fd.extension}`;
 	return vscode.Uri.joinPath(fd.parent, tempFileName);
 }
 
 export async function openFile(file:vscode.Uri) : Promise<void> {
 	const doc = await vscode.workspace.openTextDocument(file);
 	await vscode.window.showTextDocument(doc);
-}
-
-export async function callInInteractiveTerminal(command: string, terminal: vscode.Terminal): Promise<vscode.TerminalExitStatus> {
-	// wrapper function for terminal call to make it async and 
-	// return promise once terminal is closed
-	// from https://stackoverflow.com/a/72887036/1716283
-	terminal.sendText(command, false);
-	terminal.sendText("; exit");
-	return new Promise((resolve, reject) => {
-		const disposeToken = vscode.window.onDidCloseTerminal(
-			(closedTerminal) => {
-			if (closedTerminal === terminal) {
-				disposeToken.dispose();
-				if (terminal.exitStatus !== undefined) {
-					resolve(terminal.exitStatus);
-				} else {
-					reject("Terminal exited with undefined status");
-				}
-			}
-			}
-		);
-	});
 }
 
 export function gitFix(path:string) : string {
@@ -89,30 +69,62 @@ export async function delay(ms: number) {
 export async function fakeProgressUpdate(progressParameter:Progress, progress: { isDone:boolean }) : Promise<void> {
 	let rem = 100;
 	while(!progress.isDone) {
-		await delay(1000);
-		const inc = Math.floor(rem/2.5);
+		await delay(50);
+		const inc = Math.max(Math.floor(rem/20), 1);
 		rem -= inc;
-		if (inc > 1) {
+		if (rem > 0) {
 			progressParameter.report({ increment: inc});
 		}
 	}
 	return;
 }
 
-export async function decryptWithProgressBar(encryptedFile:vscode.Uri, tempFile:vscode.Uri): Promise<void> {
-	const parent = getParentUri(encryptedFile);
+export function decryptCommand(files:vscode.Uri[]) : void {
+	if (files.length === 0) {
+		noFileSelectedErrormessage();
+        return;
+	}
+
+	void decryptInPlace(files[0]);
+}
+
+export function encryptCommand(files:vscode.Uri[]) : void {
+	if (files.length === 0) {
+		noFileSelectedErrormessage();
+        return;
+	}
+
+	void encrypt(files[0]);
+}
+
+export function noFileSelectedErrormessage() : void {
+	void vscode.window.showErrorMessage('Cannot edit file directly: no file selected');
+}
+
+export async function decryptInPlace(encryptedFile:vscode.Uri) : Promise<Answer> {
+	const enc = dissectUri(encryptedFile);
+	const decryptCommand = c.decryptInPlaceCommand.replace(c.fileString, enc.fileName);
+	return await decryptWithProgressBar(encryptedFile, decryptCommand);
+}
+
+export async function decryptToTmpFile(encryptedFile:vscode.Uri, tempFile:vscode.Uri) : Promise<Answer> {
 	const enc = dissectUri(encryptedFile);
 	const temp = dissectUri(tempFile);
+	const decryptCommand = c.decryptToTmpCommand.replace(c.fileString, enc.fileName).replace(c.tempFileString, temp.fileName);
+	return await decryptWithProgressBar(encryptedFile, decryptCommand);
+}
 
-	// async decrypt with progress bar
-	const decryptTerminal = vscode.window.createTerminal({name: c.terminalDecryptName, cwd: parent.fsPath});
+export async function decryptWithProgressBar(encryptedFile:vscode.Uri, decryptCommand:string): Promise<Answer> {
+	const enc = dissectUri(encryptedFile);
+
+	let out:Answer = {stdout:'', stderr:''};
 	await vscode.window.withProgress(
 		{location: vscode.ProgressLocation.Notification, cancellable: false, title: c.decryptionString.replace(c.fileString, enc.fileName)}, 
 		async (progress) => {
 			progress.report({  increment: 0 });
 			const progressDetails = { isDone: false };
-			const decryptCommand = c.decryptionCommand.replace(c.fileString, enc.fileName).replace(c.tempFileString, temp.fileName);
-			void callInInteractiveTerminal(decryptCommand, decryptTerminal).then(() => {
+			cp.exec(decryptCommand, {cwd: enc.parent.fsPath}, (_, stdout, stderr) => {
+				out = {stdout:stdout, stderr:stderr};
 				progress.report({ increment: 100 });
 				progressDetails.isDone = true;
 				return;
@@ -120,15 +132,24 @@ export async function decryptWithProgressBar(encryptedFile:vscode.Uri, tempFile:
 			await fakeProgressUpdate(progress, progressDetails);			
 		}
 	);
-	return;
+	if (out.stderr) {
+		void vscode.window.showErrorMessage(`Error decrypting ${enc.fileName}: ${out.stderr}`);
+	}
+	return out;
 }
 
-export async function copyEncrypt(extendedTempFile:ExtendedTempFile) : Promise<string[]> {
+export function copyEncrypt(extendedTempFile:ExtendedTempFile) : Answer {
 	void fs.copyFileSync(extendedTempFile.tempFile.fsPath, extendedTempFile.originalFile.fsPath);
-	const originalFileName = dissectUri(extendedTempFile.originalFile).fileName;
-	const cwd = getParentUri(extendedTempFile.originalFile).fsPath;
-	const { stdout, stderr } = await exec(c.encryptionCommand.replace(c.fileString, originalFileName), {cwd:cwd});
-	return [stdout, stderr];
+	return encrypt(extendedTempFile.originalFile);	
+}
+
+export function encrypt(file:vscode.Uri) : Answer {
+	const fileDetails = dissectUri(file);
+	let out = {stdout:'', stderr:''};
+	cp.exec(c.encryptCommand.replace(c.fileString, fileDetails.fileName), {cwd:fileDetails.parent.fsPath}, (_, stdout, stderr) => {
+		out = {stdout:stdout, stderr:stderr};
+	});
+	return out;
 }
 
 export async function isSopsEncrypted(file:vscode.Uri) : Promise<boolean> {
@@ -157,6 +178,18 @@ export function getSopsPatternsFromFile(sopsFile:vscode.Uri) : PatternSet {
 	return [getParentUri(sopsFile).path, fileRegexes];
 }
 
-export function getTempFilePreExtension() : string {
+export function getSettingTempFilePreExtension() : string {
 	return vscode.workspace.getConfiguration().get<string>('sops-edit.tempFilePreExtension') ?? 'tmp';
+}
+
+export function getSettingOnlyUseButtons() : boolean {
+	return vscode.workspace.getConfiguration().get<boolean>('sops-edit.onlyUseButtons') ?? false;
+}
+
+export async function closeFileIfOpen(file:vscode.Uri) : Promise<void> {
+	const tabs: vscode.Tab[] = vscode.window.tabGroups.all.map(tg => tg.tabs).flat();
+	const index = tabs.findIndex(tab => tab.input instanceof vscode.TabInputText && tab.input.uri.path === file.path);
+	if (index !== -1) {
+		await vscode.window.tabGroups.close(tabs[index]);
+	}
 }