+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+# 0.0.4
+- Removed terminal use, instead using `child_process` package to handle shell commands
+- Now shows a nice vscode-native Error message when decryption/encryption cannot happen for some reason
+- Added `decrypt in-place` and `encrypt in-place` right-mouse-menu buttons
+- Added setting `Only Use Buttons`
+
 # 0.0.3
 - Bugfixes
   - No longer multiple encryption/decryption terminals opening for a single file